==> 01-Data-Preparation-01-Data-Formatting.ipynb <==
# 01 Data Preparation

Pre-preparing [ProPublica's COMPAS data](https://raw.githubusercontent.com/propublica/compas-analysis/master/compas-scores-two-years.csv) so it is compatible with Aequitas.

# >>> Import Libraries

print("Importing necessary libraries... ")

import pandas as pd

print("Libraries imported.")

# >>> Import raw COMPAS data set
# Source of script: https://github.com/dssg/aequitas/blob/master/examples/compas_data_for_aequitas.py
# I adapted it to keep first and last name
df = pd.read_csv('data/compas-scores-two-years.csv')

df = df.rename(columns={'id':'entity_id', 
                        'two_year_recid':'label_value'}) # label_value means: Did the person recidivate?

# score means: 'High', 'Medium' or 'Low' and reflects level of assessed risk of recidivism
# "High" and "Medium" are considered prediction that the defendant is charged with a felony 
# or misdemenor in the two years after administation of COMPAS assessment. "Low" is considered 
# a prediction of non-recidivism. This is based on ProPublica's interpretation of Northpointe's
# practioner guide. (Source: https://github.com/dssg/aequitas/blob/master/examples/compas_data_for_aequitas.py#L14)

df.loc[df['score_text'] == 'Low', 'score'] = str(0.0)
df.loc[df['score_text'] != 'Low', 'score'] = str(1.0)

df = df[['entity_id', 'first', 'last', 'score', 'label_value', 'race', 'sex', 'age_cat']]

df.to_csv('data/compas_for_namsor.csv', index=False)
