==> 01-Data-Preparation-03-Data-Cleaning.ipynb <==
# 01 Data Preparation

We clean the predictions made by namsor

# >>> Import Libraries

print("Importing necessary libraries... ")

from aequitas.group import Group # Aequitas, see https://github.com/dssg/aequitas/blob/master/docs/source/examples/compas_demo.ipynb
from aequitas.bias import Bias
from aequitas.fairness import Fairness
from aequitas.plotting import Plot

import pandas as pd

print("Libraries imported.")

# >>> Import COMPAS data set

print("Importing COMPAS data set... ")

df = pd.read_csv("data/compas_with_predictions.csv")

print("Data set imported. It is has {} entries and looks like this:".format(df.shape[0]))
df.head()

# >>> make sure no data is missing
df[df.isnull().any(axis=1)]

# >>> Compare values for sex as given by COMPAS and NamSor

df.sex.unique()

df.sex_pred.unique()

# make sex_pred start with a capital letter
# https://www.geeksforgeeks.org/capitalize-first-letter-of-a-column-in-pandas-dataframe/
df['sex_pred'] = df['sex_pred'].str.capitalize() 
df.sex_pred.unique()

# >>> Compare values for race as given by COMPAS and NamSor

df.race.unique()

df.race_pred.unique()

# We see the classes are not the same

def mapToCompas(namsor):
    correspondingValues = {
        'A': 'Asian',
        'B_NL': 'African-American',
        'HL': 'Hispanic',
        'W_NL': 'Caucasian'
    }
    return correspondingValues.get(namsor, namsor)

# map race_pred to one of race
# W_NL -> Caucasian
# HL -> Hispanic
# B_NL -> African-American
# A -> Asian
# nan -> Other
df['race_pred'] = df['race_pred'].apply(lambda x: mapToCompas(x)) 

df.race_pred.unique()

# Add in report:
# "Output is W_NL (white, non latino), HL (hispano latino), A (asian, non latino), B_NL (black, non latino)."
# Missing: Native American, Other
# When does NamSor return nan?
# In the COMPAS set, does african-american / caucasian exclude hispanic as well?
# there's a difference made between race and ethnicity in the census https://de.wikipedia.org/wiki/Race_(United_States_Census)

# Saving results to 'names_cat.csv'
print("Saving compas dataframe with predictions for gender and ethnicity to CSV... ")
df.to_csv("data/compas_with_predictions_cleaned.csv")
print("CSV saved!")
