==> 03-Data-Preparation-01.ipynb <==
# 03 Data Preparation

The next part should show whether NamSor performs equally well in all groups. For that we need to create two different data sets out of the data set we already have: One for each API endpoint. Then in each set we set the score column to NamSor's prediction probability and the label_value to whether NamSor's prediction was correct or not.

# >>> Import Libraries

print("Importing necessary libraries... ")

import pandas as pd

print("Libraries imported.")

# >>> Import COMPAS data set

print("Importing COMPAS data set... ")

df = pd.read_csv("data/compas_with_predictions_cleaned.csv")

print("Data set imported. It is has {} entries and looks like this:".format(df.shape[0]))
df.head()

def mapToNumber(b):
    if(b):
        return 1.0
    else:
        return 0.0

# >>> create a data set for gender
df_gender = df[['entity_id', 'first', 'last', 'sex', 'sex_pred', 'race']];
df_gender['score'] = 1 - df['sex_pred_prob']
df_gender['label_value'] = (df['sex'] == df['sex_pred'])

df_gender['label_value'] = df_gender['label_value'].apply(lambda x: mapToNumber(x)) 

df_gender

# >>> create a data set for ethnicity
df_ethnicity = df[['entity_id', 'first', 'last', 'race', 'race_pred', 'sex']];
df_ethnicity['score'] = 1 - df['race_pred_prob']
df_ethnicity['label_value'] = (df['race'] == df['race_pred'])

df_ethnicity['label_value'] = df_ethnicity['label_value'].apply(lambda x: mapToNumber(x)) 

df_ethnicity

# Saving results 
print("Saving compas dataframe with predictions for gender to CSV... ")
df_gender.to_csv("data/compas_gender_predictions.csv")
print("CSV saved!")

print("Saving compas dataframe with predictions for ethnicity to CSV... ")
df_ethnicity.to_csv("data/compas_ethnicity_predictions.csv")
print("CSV saved!")
