==> 04-Data-Exploration-01.ipynb <==
# 04 Data Exploration

Here we will first be looking at some descriptive statistics.

# >>> Import Libraries

print("Importing necessary libraries... ")

import pandas as pd

print("Libraries imported.")

# >>> Import COMPAS data set

print("Importing COMPAS data set for gender... ")

df_race = pd.read_csv("data/compas_ethnicity_predictions.csv")

print("Data set imported. It is has {} entries and looks like this:".format(df_race.shape[0]))
df_race.head()

print("Importing COMPAS data set for ethnicity... ")

df_sex = pd.read_csv("data/compas_gender_predictions.csv")

print("Data set imported. It is has {} entries and looks like this:".format(df_sex.shape[0]))
df_sex.head()

# >>> Looking at some descriptive statistics

df_sex.groupby(['sex']).describe()

df_sex.groupby(['race']).describe()

df_race.groupby(['sex']).describe()

df_race.groupby(['race']).describe()

df_race.groupby(["sex", "race"]).describe()

df_sex.groupby(["sex", "race"]).describe()
